requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
apscheduler
requests
python-dotenv
transformers
torch

